+++ src/3d-view/assets/concord.png
