+++ src/3d-view/assets/sand.png
