+++ src/3d-view/components/three-d-view.scss
