+++ src/3d-view/data/fake-data.ts
