+++ src/3d-view/index.tsx
@@ -1,0 +1,6 @@
+import * as React from "react";
+import * as ReactDOM from "react-dom";
+import { AppComponent } from "./app";
+
+ReactDOM.render(<AppComponent />, document.getElementById("app"));
+
