+++ src/3d-view/utils/translation/lang/es.json
