+++ src/common/connect-to-firestore.ts
