+++ src/common/types.ts
@@ -3,3 +3,10 @@ import { IRuntimeInitInteractive } from "@concord-consortium/lara-interactive-ap
 export interface IAuthoredState {
   firebaseApp?: string;
 }
+
+export interface AuthoringApiProps {
+  setOutput: (output: any) => void;
+  setError: (error: any) => void;
+}
+
+
