+++ src/data-table/index.tsx
@@ -1,0 +1,6 @@
+import * as React from "react";
+import * as ReactDOM from "react-dom";
+import {Test} from "./firebase-test";
+
+ReactDOM.render(<Test />, document.getElementById("app"));
+
