+++ webpack.config.js
@@ -15,13 +15,18 @@ module.exports = (env, argv) => {
     devServer: {
       static: 'dist',
       hot: true,
-      https: {
-        key: path.resolve(os.homedir(), '.localhost-ssl/localhost.key'),
-        cert: path.resolve(os.homedir(), '.localhost-ssl/localhost.pem'),
-      },
+      server: {
+        type: 'https',
+        options: {
+          cert: path.resolve(os.homedir(), '.localhost-ssl/localhost.crt'),
+          key: path.resolve(os.homedir(), '.localhost-ssl/localhost.key'),
+        }
+      }
     },
     devtool: devMode ? 'eval-cheap-module-source-map' : 'source-map',
-    entry: './src/index.tsx',
+    entry: {
+      "3dview": './src/3d-view/index.tsx'
+    },
     mode: 'development',
     output: {
       path: path.resolve(__dirname, 'dist'),
@@ -132,7 +137,7 @@ module.exports = (env, argv) => {
       }),
       new HtmlWebpackPlugin({
         filename: 'index.html',
-        template: 'src/index.html',
+        template: 'src/3d-view/index.html',
         favicon: 'src/public/favicon.ico',
       }),
       new CleanWebpackPlugin(),
